+++ NEWS.md
@@ -1,4 +1,7 @@
 # *News*
 
+# zipcodeR 0.1.1.9000 (WIP)
+Add function for obtaining congressional districts of a given ZIP code.
+
 # zipcodeR 0.1.0 (2020-09-08)
-Initial public release, to be submitted to CRAN.
+Initial public release, first version accepted by CRAN.